requirements.txt
Flask
web3
python-dotenv
ollama